pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Reader Study Web API backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "fastapi>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.4",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
    "fastapi-users>=12.1.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.1",
    "asyncpg>=0.28.0",
    "aiosqlite>=0.19.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.5",
    "httpx>=0.27.0",
]
